benchmark: submodule -> cmake ExternalProject_Add
Unnecessary download of benchmark can be avoided with CMake.
During CMake configure, benchmark is downloaded if needed and not found.
benchmark is then built and installed during CMake configure step.
We can then use find_package with REQUIRED to use benchmark.
This is cleaner and is the suggested method by google.
https://github.com/google/googletest/blob/master/googletest/README.md
Settings for benchmark build can be set in:
benchmark_download.cmake.in.
This file is configured so use of variables from main build is possible.