build(cmake): fix case where project is imported into existing CMake environement
When using something based in FetchContent like CPM (Cmake Package Manager), CMAKE_SOURCE_DIR points to the top-level of the user's project and not the root of yomm2. This results in find_or_download_package not being found. By specifying explicitly the relative path to the file, this can be avoided.